changed to tsx, renderd board, addes basic CSS.
